chore: split smt_verify into ssa/acir, start the ACIR encoder
Reorganizes smt_verify into smt_verify::ssa (everything from the previous
mod.rs: the SSA Encoder, pass/simplify equivalence checking, and their
tests) and smt_verify::acir (new), with mod.rs holding only what's shared
between them (cvc5 invocation, field helpers). Nesting each side's tests
under its own module keeps them reachable without any pub visibility
changes, the same Rust-privacy trick the SSA side already relied on.

Adds a minimal AcirEncoder covering Opcode::AssertZero, built from real
ACIR text via Circuit::from_str (acvm-repo/acir has its own parser,
parallel to Ssa::from_str). AssertZero's Expression is already a flat
polynomial (mul_terms/linear_combinations/q_c) with no expression tree to
memoize, so encoding it needs no Encoder-style name cache.

Includes two rounds of readability work on the coefficient encoding,
found by actually reading the generated snapshots: skip `* 1`/`+ 0`
wherever they're no-ops (unconditional field identities, not pattern
matching), and define `zero`/`one` named constants only when a circuit
actually uses them, instead of repeating `(as ff0 FF)` on every
AssertZero line or a ~77-digit canonical residue for small negative
coefficients.
